refactor: Reorganize ts helpers
Move helpers/index into utils.ts, and add other files which are exported in

the new index.ts.
