+++ packages/contracts/helpers/constants.ts
@@ -1,0 +1,8 @@
+/**
+ * Predeploys are Solidity contracts that are injected into the initial L2 state and provide
+ * various useful functions.
+ */
+export const bedrockPredeploys = {
+  L1_BLOCK_ATTRIBUTES: '0x4200000000000000000000000000000000000015',
+  WITHDRAWER: '0x4200000000000000000000000000000000000016',
+}
